requirements.txt
apparmor==3.1.7
asciidoc==10.2.0
beautifulsoup4==4.12.3
black==24.4.2
Brlapi==0.8.5
bs4==0.0.2
btrfsutil==6.9.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.21.2
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
html2text==2024.2.26
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
lensfun==0.3.4
LibAppArmor==3.1.7
louis==3.30.0
markdown-it-py==3.0.0
mdurl==0.1.2
meson==1.4.1
msgpack==1.0.8
mypy==1.10.1
mypy-extensions==1.0.0
netsnmp-python==1.0a1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.8
pwquality==1.4.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.26.1
Pygments==2.18.0
PyGObject==3.48.2
pynvim==0.5.0
pyparsing==3.1.2
pytest==8.2.2
python-dotenv==1.0.1
pyxdg==0.28
requests==2.32.3
rich==13.7.1
rsa==4.9
setproctitle==1.3.3
soupsieve==2.5
tabulate==0.9.0
tqdm==4.66.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2